+++ Content.Shared/_TP/DeathWhales/DeathWhaleComponent.cs
@@ -1,4 +1,4 @@
-namespace Content.Server.Deathwhale
+namespace Content.Shared._TP.DeathWhales
 {
     [RegisterComponent]
     public sealed partial class DeathWhaleComponent : Component
